Select Type -> Select Color List
Have the WineType pull up the correct Color list.  Both are sending
action payloads to Redux.

There was a problem with the return state where:

export const reducer: Reducer<BlindTastingState> =

didn't work.  But it will compile and run without error with:

export const reducer: any =

The problem seems to be caused when more than one action is given a
payload.  The problem does not occur when one action is given multiple
payloads.