Update Age to Store
Use ComponentWillReceiveProps to get Age to update the store without
falling into an infinate loop with the update causing a re-mounting that
launches a re-update and so on.  Now it checks to see if any of the
pertenent properties have changed.  If yes, it runs update.  Then when
it checks a second time, the properties should all be the same and it
will fall-through without runing updateAge.